render atoms and style mixins
