+++ libs/compiler-yew/src/code_compiler.rs
@@ -86,6 +86,9 @@ macro_rules! compile_children {
 }
 
 fn compile_component(component: &ast::Component, context: &mut Context) -> Result<()> {
+
+    let render_node = get_or_short!(component.get_render_expr(), Ok(()));
+
     compile_component_props(component, context)?;
 
     context.add_buffer("#[function_component]\n");
@@ -103,9 +106,7 @@ fn compile_component(component: &ast::Component, context: &mut Context) -> Resul
     );
     context.start_block();
     compile_render(
-        &component
-            .get_render_expr()
-            .expect("render function must exist"),
+        render_node,
         &mut context.within_component(component),
     );
     context.end_block();
