+++ libs/proto_ext/src/ast_mutate/move_node.rs
@@ -1,3 +1,5 @@
+use paperclip_proto::ast::graph_ext::Expr;
+use paperclip_proto::ast::pc::{component_body_item, Render};
 use paperclip_proto::ast_mutate::{mutation_result, ExpressionDeleted};
 use paperclip_proto::{
     ast::{
@@ -8,6 +10,7 @@ use paperclip_proto::{
 };
 
 use super::EditContext;
+use crate::ast::get_expr::GetExpr;
 use crate::{
     ast::{
         all::{MutableVisitor, VisitorResult},
@@ -177,5 +180,33 @@ impl<'a> MutableVisitor<()> for EditContext<'a, MoveNode> {
         }
 
         VisitorResult::Continue
+    }
+
+    fn visit_component(&mut self, expr: &mut paperclip_proto::ast::pc::Component) -> VisitorResult<()> {
+        if self.mutation.target_id != expr.id {
+            return VisitorResult::Continue;
+        }
+
+        let node = GetExpr::get_expr_from_graph(&self.mutation.node_id, &self.graph).expect("Node must exist");
+
+        let node = match node {
+            ExpressionWrapper::Element(node) => {
+                Some(node::Inner::Element(node).get_outer())
+            },
+            ExpressionWrapper::TextNode(node) => {
+                Some(node::Inner::Text(node).get_outer())
+            },
+            _ => {
+                None
+            }
+        };
+
+        expr.body.push(component_body_item::Inner::Render(Render {
+            id: self.get_dependency().document.as_ref().expect("Doc must exist").checksum(),
+            range: None,
+            node
+        }).get_outer());
+
+        VisitorResult::Return(())
     }
 }
